+++ Tests/EmojiKitTests/EmojiCategory+PersistedTests.swift
@@ -1,0 +1,15 @@
+//
+//  EmojiCategory+PersistedTests.swift
+//  EmojiKit
+//
+//  Created by Daniel Saidi on 2025-04-10.
+//
+
+import Testing
+
+struct Test {
+
+    @Test func testHasStandardPersistedCategories() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+}